+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 Changes since 21.08.2021:
 * created initial repository
 * created tickets directory with basic description
+* added new MovieDescription class
+* created date and dateutils classes
 * created person base class